fix package-level and baseline logger.
Due to Visdom, the logger used to print two duplicates of every log record.
By setting 'propagate=False' to the package-level logger, this does not happen anymore.

Fixed also the logger of the 'tac.agents.baseline' module: when it is run as main,
it will use the same configuration of the 'tac' logger.