fix terminate method of ControllerAgent.
Fix an issue in `ControllerAgent.terminate` method due to multithreading. Before stopping the dashboard, wait that all the incoming messages have been processed and that the message processing thread has stopped.

Otherwise, it might happen that some message handler wants to update the dashboard when it has already been stopped.